ISSUE-47: Fixed resize bug causing ghost turtle.
On certain window sizes, rounding on the original resize calculation
was causing small shifts in the turtle position. On systems where the
XOR turtle drawing works, this resulted in ghost turtles. On other
systems, there was no ghost turtle; but, post-resize drawn lines could
still be misaligned from pre-resize drawn lines.

Additionally, this commit stops the turtle from being drawn in graphics.c
if it is going to be drawn in wx in order to prevent small gaps in lines
if XOR is not supported on the system.