Fix race between GC and case object construction
Does a partial GC before constructing new case objects to ensure there are
enough free nodes to finish construction and add it to a root node
before the GC runs again.

Case objects must be marked twice so if the GC runs after a new case is
allocated but before it is added to the hash table AND a root AND the
pointer isn't found on the stack, eg because it was optimized out, then
the GC may free the cases.

I was concerned this would have a negative performance impact...

It actually improved unit test performance by 10-20%.

¯\_(ツ)_/¯